pyproject.toml
[project]
name = "aicsairtable"
version = "0.0.0"  # Do not edit manually; managed by bumpversion
description = "A package with tools to access AICS airtable data"
authors = [
    {name = "AICS Software", email="!AICS_SW@alleninstitute.org"},
]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.9.2"

dependencies = [
    "openpyxl == 3.0.10",
    "pandas == 1.5.2",
    "scipy == 1.9.3",
    "requests == 2.28.1",
    "python-dotenv ~= 0.21.0",
]

[project.optional-dependencies]
dev = [
    "black == 22.8.0",
    "build == 0.7.0",
    "bump2version ~= 1.0.1",
    "flake8 ~= 5.0.4",
    "furo == 2022.9.29",
    "isort ~= 5.9",
    "m2r2 ~= 0.3.3",
    "mypy == 0.982",
    "numpy ~= 1.23.4",
    "pre-commit ~= 2.20.0",
    "pytest ~= 7.2.0",
    "Sphinx ~= 5.3.0",
    "tifffile == 2022.10.10",
    "twine ~= 4.0.1",
    "types-python-dateutil ~= 2.8.19",
    "types-requests ~= 2.28.11",
    "wheel ~= 0.37.0",
]

[project.urls]
homepage = "https://github.com/aics-int/"
documentation = "https://aics-int.github.io/"
repository = "https://github.com/aics-int/"

[project.scripts]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = [
    "aicsairtable/tests",
]

[build-system]
# https://setuptools.pypa.io/en/latest/build_meta.html
requires = ["setuptools"]
build-backend = "setuptools.build_meta"